update books and edit-book.jsp and change methods bookServiceImpl

+++ src/main/java/com/nickstell/dao/BookDao.java
@@ -3,5 +3,5 @@ package com.nickstell.dao;
 import com.nickstell.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookDao extends JpaRepository<Book, Integer> {
+public interface BookDao extends JpaRepository<Book, Long> {
 }
